renamed hc_sync_status.dart to sync_status.dart
