+++ packages/rook_sdk_health_connect/lib/src/domain/enums/sync_status_with_data.dart
@@ -3,7 +3,7 @@ abstract interface class SyncStatusWithData<T> {}
 final class RecordsNotFound implements SyncStatusWithData<Never> {
   RecordsNotFound._();
 
-  static RecordsNotFound singleton = RecordsNotFound._();
+  static final RecordsNotFound singleton = RecordsNotFound._();
 }
 
 final class Synced<T> implements SyncStatusWithData<T> {
