+++ packages/rook_sdk_health_connect/lib/src/hc_rook_event_manager.dart
@@ -1,4 +1,4 @@
-import 'package:rook_sdk_health_connect/src/domain/enums/hc_sync_status.dart';
+import 'package:rook_sdk_health_connect/src/domain/enums/sync_status.dart';
 import 'package:rook_sdk_health_connect/src/domain/enums/sync_status_with_data.dart';
 import 'package:rook_sdk_health_connect/src/platform/rook_sdk_health_connect_platform_interface.dart';
