+++ packages/rook_sdk_health_connect/test/src/data/extension/result_sync_status_extensions_test.dart
@@ -3,7 +3,7 @@ import 'package:rook_sdk_core/rook_sdk_core.dart';
 import 'package:rook_sdk_health_connect/src/data/extension/result_boolean_extensions.dart';
 import 'package:rook_sdk_health_connect/src/data/extension/result_sync_status_extensions.dart';
 import 'package:rook_sdk_health_connect/src/data/proto/protos.pb.dart';
-import 'package:rook_sdk_health_connect/src/domain/enums/hc_sync_status.dart';
+import 'package:rook_sdk_health_connect/src/domain/enums/sync_status.dart';
 import 'package:rook_sdk_health_connect/src/domain/exception/device_not_supported_exception.dart';
 import 'package:rook_sdk_health_connect/src/domain/exception/health_connect_not_installed_exception.dart';
 import 'package:rook_sdk_health_connect/src/domain/exception/request_quota_exceeded_exception.dart';
