+++ packages/rook_sdk_health_connect/test/src/data/mapper/sync_status_mappers_test.dart
@@ -2,6 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:rook_sdk_health_connect/rook_sdk_health_connect.dart';
 import 'package:rook_sdk_health_connect/src/data/mapper/sync_status_mappers.dart';
 import 'package:rook_sdk_health_connect/src/data/proto/protos.pbenum.dart';
+import 'package:rook_sdk_health_connect/src/domain/enums/sync_status.dart';
 
 void main() {
   group('SyncStatusProto to SyncStatus mappers', () {
